# docs/swagger.yaml
definitions:
...
  entity.HomevisitSales:
    properties:
      available:
        type: integer
      createdAt:
        type: string
      deposit:
        type: integer
      endTime:
        type: string
      id:
        type: integer
      merchantId:
        type: integer
      startTime:
        type: string
      total:
        type: integer
      updatedAt:
        type: string
    type: object
...
  entity.Role:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.ReqBatchCreateHomevisitSales:
    properties:
      request:
        items:
          $ref: '#/definitions/model.ReqCreateHomevisitSales'
        type: array
    type: object
...
  model.ReqCreateHomeDetailVisitSales:
    properties:
      endTime:
        type: string
      numberOfFoodtruck:
        type: integer
      startTime:
        type: string
    required:
    - endTime
    - numberOfFoodtruck
    - startTime
    type: object
  model.ReqCreateHomevisitSales:
    properties:
      date:
        type: string
      deposit:
        type: integer
      summary:
        items:
          $ref: '#/definitions/model.ReqCreateHomeDetailVisitSales'
        type: array
    required:
    - date
    - deposit
    type: object
...
  model.ReqRoleCreate:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  model.ReqUpdateHomeDetailVisitSales:
    properties:
      endTime:
        type: string
      id:
        type: integer
      numberOfFoodtruck:
        type: integer
      startTime:
        type: string
    required:
    - endTime
    - id
    - numberOfFoodtruck
    - startTime
    type: object
  model.ReqUpdateHomevisitSales:
    properties:
      date:
        type: string
      deposit:
        type: integer
      summary:
        items:
          $ref: '#/definitions/model.ReqUpdateHomeDetailVisitSales'
        type: array
    required:
    - date
    - deposit
    - summary
    type: object
...
  model.ResHomeVisitDetails:
    properties:
      endTime:
        type: string
      id:
        type: integer
      numberOfFoodtruck:
        type: integer
      startTime:
        type: string
    type: object
  model.ResHomeVisitGetInfo:
    properties:
      date:
        type: string
      deposit:
        type: integer
      summary:
        items:
          $ref: '#/definitions/model.ResHomeVisitDetails'
        type: array
    type: object
...
  model.ResponseErrors:
    properties:
      error:
        $ref: '#/definitions/model.errorResponse'
    type: object
  model.ResponseSuccess:
    properties:
      data:
        type: object
    type: object
...
  model.errorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /homevisit:
    get:
      operationId: GetAllHV
      parameters:
      - default: "10"
        description: ' '
        in: query
        name: limit
        type: string
      - default: "1"
        description: ' '
        in: query
        name: page
        type: string
      - description: 'e.g.: id,desc / id,asc'
        in: query
        name: sort
        type: string
      responses:
        "200":
          description: model.Pagination
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
      security:
      - Token: []
      summary: Get All Home Visit Pagination (permission = admin)
      tags:
      - Home Visit
    post:
      operationId: CreateHV
      parameters:
      - description: all fields mandatory
        in: body
        name: homeVisit
        required: true
        schema:
          $ref: '#/definitions/model.ReqCreateHomevisitSales'
      responses:
        "200":
          description: 'data: []entity.HomevisitSales'
          schema:
            items:
              $ref: '#/definitions/entity.HomevisitSales'
            type: array
      security:
      - Token: []
      summary: Create New Home Visit (permission = admin)
      tags:
      - Home Visit
    put:
      operationId: UpdateHomeVisit
      parameters:
      - description: all fields mandatory
        in: body
        name: homeVisit
        required: true
        schema:
          $ref: '#/definitions/model.ReqUpdateHomevisitSales'
      responses:
        "200":
          description: 'data: []entity.HomevisitSales'
          schema:
            items:
              $ref: '#/definitions/entity.HomevisitSales'
            type: array
      security:
      - Token: []
      summary: Update Home Visit (permission = admin)
      tags:
      - Home Visit
  /homevisit/batch:
    post:
      operationId: BatchCreateHV
      parameters:
      - description: all fields mandatory
        in: body
        name: batchHomeVisit
        required: true
        schema:
          $ref: '#/definitions/model.ReqBatchCreateHomevisitSales'
      responses:
        "200":
          description: 'data: []entity.HomevisitSales'
          schema:
            items:
              $ref: '#/definitions/entity.HomevisitSales'
            type: array
      security:
      - Token: []
      summary: Batch Create New Home Visit (permission = admin)
      tags:
      - Home Visit
  /homevisit/deletebydate/{date}:
    delete:
      operationId: DeleteHomeVisit
      parameters:
      - description: yyyy-mm-dd
        in: path
        name: date
        required: true
        type: string
      responses:
        "200":
          description: 'message: "successfully deleted home visit!" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: 'message: "Failed to Delete Menu'
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Delete Home Visit (permission = admin)
      tags:
      - Home Visit
  /homevisit/deletebyid/{id}:
    delete:
      operationId: DeleteHomeVisitByID
      parameters:
      - description: number
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 'message: "successfully deleted home visit!" '
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "500":
          description: 'message: "Failed to Delete Menu'
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Delete Home Visit By ID (permission = admin)
      tags:
      - Home Visit
  /homevisit/info/{date}:
    get:
      operationId: GetInfoHV
      parameters:
      - description: yyyy-mm-dd
        in: path
        name: date
        required: true
        type: string
      responses:
        "200":
          description: 'data: model.ResHomeVisitGetInfo'
          schema:
            $ref: '#/definitions/model.ResHomeVisitGetInfo'
      security:
      - Token: []
      summary: Get Info Home Visit (permission = admin)
      tags:
      - Home Visit
...
  /log-merchant:
    get:
      operationId: GetAllLogMerchant
      parameters:
      - default: "10"
        description: ' '
        in: query
        name: limit
        type: string
      - default: "1"
        description: ' '
        in: query
        name: page
        type: string
      - description: 'e.g.: id,desc / id,asc'
        in: query
        name: sort
        type: string
      responses:
        "200":
          description: model.Pagination
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
      security:
      - Token: []
      summary: Get All Log Merchant Pagination (permission = admin)
      tags:
      - Log Merchant
...
  /role:
    get:
      operationId: GetAllRole
      responses:
        "200":
          description: 'data: []entity.Role '
          schema:
            items:
              $ref: '#/definitions/entity.Role'
            type: array
      security:
      - Token: []
      summary: Get All Role (permission = superadmin)
      tags:
      - Master Role
    post:
      operationId: CreateRole
      parameters:
      - description: role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ReqRoleCreate'
      responses:
        "200":
          description: 'data: entity.Role, message: "New Role has created" '
          schema:
            $ref: '#/definitions/entity.Role'
        "422":
          description: 'code: 422, message: "Invalid request" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
        "500":
          description: 'code: 500, message: "Create Role Failed" '
          schema:
            $ref: '#/definitions/model.ResponseErrors'
      security:
      - Token: []
      summary: Create new Role (permission = superadmin)
      tags:
      - Master Role
  /role/exclude:
    get:
      operationId: GetAllExclude
      responses:
        "200":
          description: 'data: []entity.Role '
          schema:
            items:
              $ref: '#/definitions/entity.Role'
            type: array
      security:
      - Token: []
      summary: Get All Role Exclude Foodtruck (permission = superadmin)
      tags:
      - Master Role
...
swagger: "2.0"